chore: move PluginUtilsService out of the tasks package so it could be removed soon
