+++ core/src/test/java/io/kestra/core/tasks/PluginUtilsServiceTest.java
@@ -1,6 +1,7 @@
 package io.kestra.core.tasks;
 
 import io.kestra.core.exceptions.IllegalVariableEvaluationException;
+import io.kestra.core.models.tasks.runners.PluginUtilsService;
 import io.kestra.core.runners.RunContext;
 import io.micronaut.context.ApplicationContext;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
